+++ packages/empty/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @siakit/empty
 
+## 0.0.6
+
+### Patch Changes
+
+- fix empty component
+
+## 0.0.5
+
+### Patch Changes
+
+- fix empty component
+
 ## 0.0.4
 
 ### Patch Changes